// modules/cip/types/cip.ts
export type CipPassengerType = "Adult" | "Child" | "Infant" | "Accompanying";
// ...
export interface CipGetReserveByIdResponse {
  flightNumber: string;
  airline: string;
  flightTime: string;
  username: string;
  status:  "Undefined" | "Registered" | "Pending" | "Issued" | "Canceled" | "WebServiceCancel" | "PaymentSuccessful" | "WebServiceUnsuccessful" | "PriceChange" | "Unavailable" | "Refunded" | "Voided" | "InProgress" | "PaidBack" | "RefundInProgress" | "Changed" | "OnCredit" | "ContactProvider" | "UnConfirmed" | "ReceivedAdvance" | "ExtraReceiving" ;
  reserver: {
    firstName: string;
    lastName: string;
    phoneNumber: string;
    email: string;
    userName: string;
    gender: boolean;
  };
  airport: {
    name: string;
    city: {
      name: string;
      code: string;
    };
    picture: {
      path: string;
      altAttribute: string;
      titleAttribute: string;
    },
    id: number;
  };
  passengers: [
    {
      gender: boolean;
      firstName: string;
      lastName: string;
      passengerType: CipPassengerType;
      passportNumber: string;
      nationalId: string;
      nationality: string;
      birthday: string,
      service: {
        items: [
          {
            name: string;
            picture: {
              path: string;
              altAttribute: string;
              titleAttribute: string;
            };
            id: number
          }
        ]
      },
      id: number;
    }
  ],
  transport: {
    items: [
      {
        name: string;
        address: string;
        count: number;
        salePrice: number;
        picture: {
          path: string;
          altAttribute: string;
          titleAttribute: string;
        };
      }
    ]
  };
  service: {
    items: [
      {
        count: number;
        name: string;
        type: string;
        salePrice: number;
        picture: {
          path: string;
          altAttribute: string;
          titleAttribute: string;
        };
      }
    ];
  };
  id: number;
}